requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
deep-translator==1.11.4
dj-database-url==2.3.0
Django==5.1.6
django-ckeditor==6.7.2
django-filter==24.3
django-js-asset==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg==3.2.4
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
